Rename transformers and flatten package structure
